Auto-detect the cjdns port for iptables
The tradeoff to specifying the variable manually is that it will not be
detected on the initial playbook run. This is due to the purposeful
separation of the common and cjdns roles, so I did not put a handler
into place to trigger a refresh once the port was known.