Fix python PATH issue when building libuv dependency
This is an ugly hack to fix the build issues that occur when running the
./do script normally. There was a commit [1] to fix the python2 PATH
issue (since the `python` binary in Arch defaults to Python 3), however,
it still breaks when building the libuv dependency unless you hard-code
the binary name.

I'm hoping to figure out a cleaner workaround, but this gets cjdns
compiling again (and is also what prurigro has done with the AUR
PKGBUILD).

[1] https://github.com/cjdelisle/cjdns/commit/ee58fb440f00e6815b6fcab4d28124f06d2e1d33